Split the class into separate files.
