+++ MetaphysicsIndustries.Solus.csproj
@@ -59,6 +59,62 @@
     <Compile Include="Compiler\CompiledExpression.cs" />
     <Compile Include="Compiler\IILGenerator.cs" />
     <Compile Include="Compiler\ILCompiler.cs" />
+    <Compile Include="Compiler\ILCompiler.Expressions.ComponentAccess.cs" />
+    <Compile Include="Compiler\ILCompiler.Expressions.cs" />
+    <Compile Include="Compiler\ILCompiler.Expressions.FunctionCall.cs" />
+    <Compile Include="Compiler\ILCompiler.Expressions.Literal.cs" />
+    <Compile Include="Compiler\ILCompiler.Expressions.MatrixExpression.cs" />
+    <Compile Include="Compiler\ILCompiler.Expressions.VariableAccess.cs" />
+    <Compile Include="Compiler\ILCompiler.Expressions.VectorExpression.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.AbsoluteValueFunction.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.AdditionOperation.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.ArccosecantFunction.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.ArccosineFunction.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.ArccotangentFunction.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.ArcsecantFunction.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.ArcsineFunction.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.Arctangent2Function.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.ArctangentFunction.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.BitwiseAndOperation.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.BitwiseOrOperation.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.CeilingFunction.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.CosecantFunction.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.CosineFunction.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.CotangentFunction.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.DistFunction.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.DistSqFunction.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.DivisionOperation.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.EqualComparisonOperation.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.ExponentOperation.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.FactorialFunction.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.FloorFunction.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.GreaterThanComparisonOperation.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.GreaterThanOrEqualComparisonOperation.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.LessThanComparisonOperation.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.LessThanOrEqualComparisonOperation.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.LoadImageFunction.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.Log10Function.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.Log2Function.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.LogarithmFunction.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.LogicalAndOperation.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.LogicalOrOperation.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.MaximumFiniteFunction.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.MaximumFunction.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.MinimumFiniteFunction.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.MinimumFunction.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.ModularDivision.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.MultiplicationOperation.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.NaturalLogarithmFunction.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.NegationOperation.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.NotEqualComparisonOperation.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.SecantFunction.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.SineFunction.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.SizeFunction.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.TangentFunction.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.UnitStepFunction.cs" />
+    <Compile Include="Compiler\ILCompiler.Functions.UserDefinedFunction.cs" />
+    <Compile Include="Compiler\ILCompiler.StoreOp.cs" />
     <Compile Include="Compiler\IlExpressions\AddIlExpression.cs" />
     <Compile Include="Compiler\IlExpressions\AndIlExpression.cs" />
     <Compile Include="Compiler\IlExpressions\BoxIlExpression.cs" />
